fix(config): config cahnces are now probabilites

+++ tsg/types.py
@@ -37,7 +37,7 @@ class TSGConfig:
     resolution_h: int
     world_width: int
     world_height: int
-    stuff_chance: int
-    gack_chance: int
+    stuff_chance: float
+    gack_chance: float
     gack_size: int
-    thing_chance: int
+    thing_chance: float
